Fix Errors

+++ ffuf.sh
@@ -6,5 +6,7 @@ if [ ! -d /opt/ffuf ]; then
     git clone https://github.com/ffuf/ffuf
     cd /opt/ffuf;go get;go build
     cd -
-    sudo ln -s /opt/ffuf/ffuf /usr/bin/ffuf  
+    if [ ! -f /usr/bin/ffuf ] then;
+        sudo ln -s /opt/ffuf/ffuf /usr/bin/ffuf  
+    fi
 fi
